requirements.txt

httpx
fastmcp
python-dotenv
aiofiles
pandas
python-docx
lxml
openpyxl
fastapi
google-generativeai
